Don't load barchart pixel images unless in print mode.
Adds a transparent 1x1 gif that is loaded once and used as an image in every bar.
This makes it possible to use a css3 attribute selector to assign css only to bars.
This is a hack that will be removed when moving to a new chart system.

Fix some css bugs and remove some unused css.